requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
python-docx>=1.1.0
spacy>=3.7.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
openpyxl>=3.1.0
reportlab>=4.0.0

